+++ kinto_signer/serializer.py
@@ -1,3 +1,4 @@
+import re
 import json
 import operator
 
@@ -8,4 +9,10 @@ def canonical_json(records, last_modified):
 
     payload = {'data': records, 'last_modified': '%s' % last_modified}
 
-    return json.dumps(payload, sort_keys=True, separators=(',', ':'))
+    dump = json.dumps(payload, sort_keys=True, separators=(',', ':'))
+
+    # Fix scientific notations of Python JSON to conform with ECMAScript v6
+    # 9.30258908e-07 --> 9.30258908e-7
+    dump = dump.replace("e-0", "e-")
+
+    return dump